rename some more files
