+++ frontend/src/main.tsx
@@ -20,7 +20,7 @@ import "./index.css"
 import VideoOptions, {videoOptionsLoader} from "./routes/VideoOptions"
 import Progress from "./routes/CreateVideoPage"
 import PreviewClips from "./routes/clips/ClipPreviewPage"
-import ListVideos from "./routes/library/videos"
+import ListVideos from "./routes/library/ListVideos"
 import CreateLayout from "./routes/Layout"
 import Layout from "./components/Layout"
 import Music from "./routes/music/MusicPage"
@@ -35,19 +35,19 @@ import {
 } from "./routes/loaders"
 import {DndProvider} from "react-dnd"
 import {HTML5Backend} from "react-dnd-html5-backend"
-import MarkersPage from "./routes/library/markers"
+import MarkersPage from "./routes/library/SelectMarkers"
 import {resetForm} from "./routes/actions"
-import AddVideosPage from "./routes/library/add"
+import AddVideosPage from "./routes/library/add/AddVideosPaage"
 import useNotification from "./hooks/useNotification"
 import {FormStage} from "./types/form-state"
-import DownloadVideosPage from "./routes/library/add/download"
-import SelectVideos from "./routes/library/add/folder"
-import AddStashVideoPage from "./routes/library/add/stash"
+import DownloadVideosPage from "./routes/library/add/DownloadVideosPage"
+import SelectVideos from "./routes/library/add/AddLocalVideosPage"
+import AddStashVideoPage from "./routes/library/add/AddStashVideoPage"
 import {ConfigProvider} from "./hooks/useConfig"
 import StashConfigPage from "./routes/StashConfig"
 import FunscriptPage from "./routes/FunscriptPage"
 import DownloadVideoPage from "./routes/DownloadFinishedVideo"
-import SelectVideosPage from "./routes/library/select-videos"
+import SelectVideosPage from "./routes/library/SelectVideos"
 import DownloadMusic from "./routes/music/DownloadMusic"
 import UploadMusic from "./routes/music/UploadMusic"
 import ReorderSongs from "./routes/music/ReorderSongs"
